changes to localStorage
