+++ src/modules/removeFromStorage.test.js
@@ -1,20 +1,20 @@
 /** * @jest-environment jsdom */
 
-import Storage from "./storage";
-import removeFromStorage from '../index.js';
+import updateid from './updateid.js';
+import removeFromStorage from './removeFromStorage.js';
 
 window.localStorage = Storage.prototype;
 
 const newTasks = [
   {
-    description: 'Vanessa',
+    description: 'Mátyás',
     id: 1,
     completed: false,
     editable: false,
   },
 
   {
-    description: 'Mátyás',
+    description: 'Vanessa',
     id: 2,
     completed: false,
     editable: false,
@@ -25,17 +25,17 @@ localStorage.setItem('tasks', JSON.stringify(newTasks));
 
 describe('check the localStorage', () => {
   test('Remove from storage', () => {
-    removeFromStorage(1);
-    const task = [
+    const task1 = [
       {
         description: 'Mátyás',
-        id: 2,
+        id: 1,
         completed: false,
         editable: false
       }
     ];
-    const storage = JSON.parse(localStorage.getItem('tasks')) || [];
-    expect(true).toBe(true)
+    removeFromStorage(2);
+    const dataStorage = JSON.parse(localStorage.getItem('tasks')) || [];
+    expect(dataStorage).toEqual(task1)
   });
 });
